+++ src/components/tabs/details-tab/details-tab.tsx
